feat: generateTest(), input to double, refact
Add and complete function generateTest().
Coordinates from user input parse to double.
Add readDouble in Input class.
Add new menu option.
Fix pathToString.

+++ planetarium/solarsystem/Planet.java
@@ -103,7 +103,7 @@ public class Planet extends CelestialBody {
      * @param moonMass The new moon's mass.
      * @see Moon
      */
-    public void addNewMoon(long relativeX, long relativeY, long moonMass) {
+    public void addNewMoon(double relativeX, double relativeY, long moonMass) {
         addNewMoon(new Position(relativeX, relativeY), moonMass);
     }
